pyproject.toml
[project]
name = "agent-api"
version = "0.1.0"
requires-python = ">=3.11"
readme = "README.md"
authors = [{ name = "Agno", email = "hello@agno.com" }]

dependencies = [
   "alembic",
  "agno",
  "agno[aws]",
  "agno-docker",
  "aiofiles",
  "beautifulsoup4",
  "duckduckgo-search",
  "fastapi[standard]",
  "mypy",
  "nest_asyncio",
  "openai",
  "pgvector",
  "pillow",
  "psycopg[binary]",
  "pypdf",
  "pytest",
  "python-docx",
  "ruff",
  "sqlalchemy",
  "tiktoken",
  "typer",
  "types-beautifulsoup4",
  "types-Pillow",
  "google-genai",
  "google-generativeai",
  "google-search-results",
  "tavily-python",
  "newspaper4k",
  "lxml_html_clean",
  "exa_py",
  "googlesearch-python",
  "pycountry",
  "mistralai",
  "mcp",
  "groq",
  "huggingface_hub",
  "tantivy",
  "inquirer",
  "jinaai",
  "bs4",
  "crawl4ai",
  "playwright",
  "scholarly",
  "googlemaps",
  "google-maps-places",
  "qdrant-client",
  "redis",
  "slowapi",
  "qdrant-client[fastembed]",
  "numpy",
  "fastembed",
  "cohere",
  "infinity_client"


]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]

# Change this value to use a different directory for the phidata workspace.
# [tool.phidata]
# workspace = "workspace"

[tool.ruff]
line-length = 110
exclude = ["aienv*", ".venv*"]
[tool.ruff.lint.per-file-ignores]
# Ignore `F401` (import violations) in all `__init__.py` files
"__init__.py" = ["F401"]

[tool.mypy]
check_untyped_defs = true
no_implicit_optional = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
exclude = ["aienv*", ".venv*"]

[[tool.mypy.overrides]]
module = ["pgvector.*", "setuptools.*", "agno.*", "agno.docker.*", "agno.aws.*"]
ignore_missing_imports = true

[tool.uv.pip]
no-annotate = true

[tool.pytest.ini_options]
log_cli = true
